Abstruct  DebeziumConnector
