+++ connectors/rocketmq-connect-debezium/rocketmq-connect-debezium-mongodb/src/main/java/org/apache/rocketmq/connect/debezium/mongodb/DebeziumMongoDBConnector.java
@@ -15,7 +15,7 @@
  *  limitations under the License.
  */
 
-package org.apache.pulsar.io.debezium.mongodb;
+package org.apache.rocketmq.connect.debezium.mongodb;
 
 import io.openmessaging.KeyValue;
 import io.openmessaging.connector.api.component.task.Task;
