+++ connectors/rocketmq-connect-debezium/rocketmq-connect-debezium-mysql/src/main/java/org/apache/rocketmq/connect/debezium/mysql/DebeziumMysqlConnector.java
@@ -19,34 +19,17 @@ package org.apache.rocketmq.connect.debezium.mysql;
 
 import io.openmessaging.KeyValue;
 import io.openmessaging.connector.api.component.task.Task;
-import io.openmessaging.connector.api.component.task.sink.SinkConnector;
+import org.apache.rocketmq.connect.debezium.DebeziumConnector;
 
-import java.util.Collections;
-import java.util.List;
 
 
 /**
  * debezium mysql connector
  */
-public class DebeziumMysqlConnector extends SinkConnector {
-
-    private KeyValue config;
-
-    /**
-     * Returns a set of configurations for Tasks based on the current configuration,
-     * producing at most count configurations.
-     *
-     * @param maxTasks maximum number of configurations to generate
-     * @return configurations for Tasks
-     */
-    @Override
-    public List<KeyValue> taskConfigs(int maxTasks) {
-        return Collections.singletonList(config);
-    }
+public class DebeziumMysqlConnector extends DebeziumConnector {
 
     /**
      * Return the current connector class
-     *
      * @return task implement class
      */
     @Override
@@ -62,39 +45,5 @@ public class DebeziumMysqlConnector extends SinkConnector {
     @Override
     public void validate(KeyValue config) {
         // do nothing
-    }
-
-    /**
-     * Init the component
-     *
-     * @param config component config
-     */
-    @Override
-    public void init(KeyValue config) {
-        this.config = config;
-    }
-
-    /**
-     * Stop the component.
-     */
-    @Override
-    public void stop() {
-
-    }
-
-    /**
-     * Pause the connector.
-     */
-    @Override
-    public void pause() {
-
-    }
-
-    /**
-     * Resume the connector.
-     */
-    @Override
-    public void resume() {
-
     }
 }
