+++ connectors/rocketmq-connect-debezium/rocketmq-connect-debezium-oracle/src/main/resources/debezium-oracle-source-config.yaml
@@ -17,22 +17,3 @@
 # under the License.
 #
 
-tenant: "public"
-namespace: "default"
-name: "debezium-oracle-source"
-topicName: "debezium-oracle-topic"
-archive: "connectors/pulsar-io-debezium-oracle-2.9.0-SNAPSHOT.nar"
-
-parallelism: 1
-
-configs:
-  ## config for Oracle XE, docker image: https://github.com/MaksymBilenko/docker-oracle-12c
-  ## docker run -d -p 1521:1521 quay.io/maksymbilenko/oracle-12c
-  database.hostname: "localhost"
-  database.port: "1521"
-  database.user: "sysdba"
-  database.password: "oracle"
-  database.dbname: "XE"
-  database.server.name: "XE"
-
-  database.history.pulsar.service.url: "pulsar://127.0.0.1:6650"
