+++ connectors/rocketmq-connect-debezium/rocketmq-connect-debezium-postgresql/src/main/java/org/apache/rocketmq/connect/debezium/postgres/DebeziumPostgresConnector.java
@@ -19,34 +19,16 @@ package org.apache.rocketmq.connect.debezium.postgres;
 
 import io.openmessaging.KeyValue;
 import io.openmessaging.connector.api.component.task.Task;
-import io.openmessaging.connector.api.component.task.sink.SinkConnector;
-
-import java.util.Collections;
-import java.util.List;
+import org.apache.rocketmq.connect.debezium.DebeziumConnector;
 
 
 /**
  * debezium postgresql connector
  */
-public class DebeziumPostgresConnector extends SinkConnector {
-
-    private KeyValue config;
-
-    /**
-     * Returns a set of configurations for Tasks based on the current configuration,
-     * producing at most count configurations.
-     *
-     * @param maxTasks maximum number of configurations to generate
-     * @return configurations for Tasks
-     */
-    @Override
-    public List<KeyValue> taskConfigs(int maxTasks) {
-        return Collections.singletonList(config);
-    }
+public class DebeziumPostgresConnector extends DebeziumConnector {
 
     /**
      * Return the current connector class
-     *
      * @return task implement class
      */
     @Override
@@ -56,45 +38,10 @@ public class DebeziumPostgresConnector extends SinkConnector {
 
     /**
      * Should invoke before start the connector.
-     *
      * @param config component config
      */
     @Override
     public void validate(KeyValue config) {
         // do nothing
-    }
-
-    /**
-     * Init the component
-     *
-     * @param config component config
-     */
-    @Override
-    public void init(KeyValue config) {
-        this.config = config;
-    }
-
-    /**
-     * Stop the component.
-     */
-    @Override
-    public void stop() {
-
-    }
-
-    /**
-     * Pause the connector.
-     */
-    @Override
-    public void pause() {
-
-    }
-
-    /**
-     * Resume the connector.
-     */
-    @Override
-    public void resume() {
-
     }
 }
