+++ connectors/rocketmq-connect-debezium/rocketmq-connect-debezium-sqlserver/src/main/resources/debezium-sqlserver-source-config.yaml
@@ -13,21 +13,7 @@
 # software distributed under the License is distributed on an
 # "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 # KIND, either express or implied.  See the License for the
-# specific language governing permissions and limitations
+# speercific language governing permissions and limitations
 # under the License.
 #
-
-
-name: "debezium-mssql-source"
-topicName: "debezium-mssql-topic"
-
-parallelism: 1
-
-configs:
-  database.hostname: "localhost"
-  database.port: "1521"
-  database.user: "sa"
-  database.password: "MyP@ssword1"
-  database.dbname: "MyDB"
-  database.server.name: "mssql"
